Refactor schemas package
