+++ src/endpoints/v2/motorized_mobility.py
@@ -8,16 +8,20 @@ from src.crud.crud_nearby_station_access import CRUDNearbyStationAccess
 from src.db.session import AsyncSession
 from src.endpoints.deps import get_db, get_http_client, get_user_id
 from src.schemas.job import JobType
-from src.schemas.motorized_mobility import (
-    IIsochroneCar,
+from src.schemas.isochrone import (
     IIsochronePT,
+    IIsochroneCar,
+    request_examples_isochrone_pt,
+    request_examples_isochrone_car,
+)
+from src.schemas.oev_gueteklasse import (
     IOevGueteklasse,
     request_example_oev_gueteklasse,
-    request_examples_isochrone_car,
-    request_examples_isochrone_pt,
-    request_example_nearby_station_access,
-    ITripCountStation,
+)
+from src.schemas.trip_count_station import ITripCountStation
+from src.schemas.nearby_station_access import (
     INearbyStationAccess,
+    request_example_nearby_station_access,
 )
 from src.schemas.toolbox_base import IToolResponse
 from src.schemas.toolbox_base import CommonToolParams
