+++ src/schemas/isochrone.py
@@ -4,8 +4,18 @@ from uuid import UUID
 from pydantic import BaseModel, Field, validator
 
 from src.schemas.layer import ToolType
-from src.schemas.toolbox_base import IsochroneStartingPointsBase, check_starting_points, input_layer_type_point
 from src.schemas.colors import ColorRangeType
+
+from typing import List, Optional
+from src.schemas.toolbox_base import (
+    IsochroneStartingPointsBase,
+    PTTimeWindow,
+    input_layer_type_point,
+    check_starting_points,
+)
+
+
+"""Isochrone starting point validators."""
 
 class IsochroneStartingPointsActiveMobility(IsochroneStartingPointsBase):
     """Model for the active mobility isochrone starting points."""
@@ -14,7 +24,16 @@ class IsochroneStartingPointsActiveMobility(IsochroneStartingPointsBase):
     check_starting_points = check_starting_points(1000)
 
 
-class RoutingActiveMobilityType(str, Enum):
+class IsochroneStartingPointsMotorizedMobility(IsochroneStartingPointsBase):
+    """Model for the active mobility isochrone starting points."""
+
+    # Check that the starting points for motorized mobility is 1
+    check_starting_points = check_starting_points(1)
+
+
+"""Isochrone routing mode schemas."""
+
+class IsochroneRoutingModeActiveMobility(str, Enum):
     """Routing active mobility type schema."""
 
     walking = "walking"
@@ -22,7 +41,62 @@ class RoutingActiveMobilityType(str, Enum):
     pedelec = "pedelec"
 
 
-class TravelTimeCostActiveMobility(BaseModel):
+class IsochroneRoutingModePT(str, Enum):
+    """Routing public transport mode schema."""
+
+    bus = "bus"
+    tram = "tram"
+    rail = "rail"
+    subway = "subway"
+    ferry = "ferry"
+    cable_car = "cable_car"
+    gondola = "gondola"
+    funicular = "funicular"
+
+
+class IsochroneRoutingEgressModePT(str, Enum):
+    """Routing public transport egress mode schema."""
+    walk = "walk"
+    bicycle = "bicycle"
+
+
+class IsochroneRoutingAccessModePT(str, Enum):
+    """Routing public transport access mode schema."""
+
+    walk = "walk"
+    bicycle = "bicycle"
+    car = "car"
+
+
+class IsochroneRoutingModeConfigPT(BaseModel):
+    """Routing public transport type schema."""
+
+    mode: List[IsochroneRoutingModePT] = Field(
+        ...,
+        title="Mode",
+        description="The mode of the public transport.",
+    )
+    egress_mode: IsochroneRoutingEgressModePT = Field(
+        ...,
+        title="Egress Mode",
+        description="The egress mode of the public transport.",
+    )
+    access_mode: IsochroneRoutingAccessModePT = Field(
+        ...,
+        title="Access Mode",
+        description="The access mode of the public transport.",
+    )
+
+
+class IsochroneRoutingTypeCar(str, Enum):
+    """Routing car type schema."""
+
+    car_peak = "car_peak"
+
+
+"""Isochrone travel cost schemas."""
+
+class IsochroneTravelTimeCostActiveMobility(BaseModel):
     """Travel time cost schema."""
 
     max_traveltime: int = Field(
@@ -56,7 +130,7 @@ class TravelTimeCostActiveMobility(BaseModel):
 
 
 # TODO: Check how to treat miles
-class TravelDistanceCostActiveMobility(BaseModel):
+class IsochroneTravelDistanceCostActiveMobility(BaseModel):
     """Travel distance cost schema."""
 
     max_distance: int = Field(
@@ -82,13 +156,60 @@ class TravelDistanceCostActiveMobility(BaseModel):
         return v
 
 
-class IsochroneType(str, Enum):
-    """Isochrone type schema."""
+class IsochroneTravelTimeCostMotorizedMobility(BaseModel):
+    """Travel time cost schema."""
+
+    max_traveltime: int = Field(
+        ...,
+        title="Max Travel Time",
+        description="The maximum travel time in minutes.",
+        ge=1,
+        le=60,
+    )
+    steps: int = Field(
+        ...,
+        title="Steps",
+        description="The number of steps.",
+    )
+
+
+"""Isochrone decay function schemas."""
+
+class IsochroneDecayFunctionTypePT(Enum):
+    LOGISTIC = "logistic"
+    LINEAR = "linear"
+    EXPONENTIAL = "exponential"
+    STEP = "step"
+
+
+class IsochroneDecayFunctionPT(BaseModel):
+    type: Optional[IsochroneDecayFunctionTypePT] = Field(
+        IsochroneDecayFunctionTypePT.LOGISTIC, description="Decay function type"
+    )
+    standard_deviation_minutes: Optional[int] = Field(
+        12, description="Standard deviation in minutes"
+    )
+    width_minutes: Optional[int] = Field(10, description="Width in minutes")
+
+
+"""Isochrone type schemas."""
+
+class IsochroneTypeActiveMobility(str, Enum):
+    """Isochrone type schema for active mobility."""
 
     polygon = "polygon"
     network = "network"
     rectangular_grid = "rectangular_grid"
 
+
+class IsochroneTypePT(str, Enum):
+    """Isochrone type schema for public transport."""
+
+    polygon = "polygon"
+    rectangular_grid = "rectangular_grid"
+
+
+"""User-configured isochrone payload schemas."""
 
 class IIsochroneActiveMobility(BaseModel):
     """Model for the active mobility isochrone"""
@@ -98,12 +219,12 @@ class IIsochroneActiveMobility(BaseModel):
         title="Starting Points",
         description="The starting points of the isochrone.",
     )
-    routing_type: RoutingActiveMobilityType = Field(
+    routing_type: IsochroneRoutingModeActiveMobility = Field(
         ...,
         title="Routing Type",
         description="The routing type of the isochrone.",
     )
-    travel_cost: TravelTimeCostActiveMobility | TravelDistanceCostActiveMobility = (
+    travel_cost: IsochroneTravelTimeCostActiveMobility | IsochroneTravelDistanceCostActiveMobility = (
         Field(
             ...,
             title="Travel Cost",
@@ -115,7 +236,7 @@ class IIsochroneActiveMobility(BaseModel):
         title="Scenario ID",
         description="The ID of the scenario that is used for the routing.",
     )
-    isochrone_type: IsochroneType = Field(
+    isochrone_type: IsochroneTypeActiveMobility = Field(
         ...,
         title="Return Type",
         description="The return type of the isochrone.",
@@ -149,6 +270,117 @@ class IIsochroneActiveMobility(BaseModel):
         }
 
 
+class IIsochronePT(BaseModel):
+    """Model for the public transport isochrone"""
+
+    starting_points: IsochroneStartingPointsMotorizedMobility = Field(
+        ...,
+        title="Starting Points",
+        description="The starting points of the isochrone.",
+    )
+    routing_type: IsochroneRoutingModeConfigPT = Field(
+        ...,
+        title="Routing Type",
+        description="The routing type of the isochrone.",
+    )
+    travel_cost: IsochroneTravelTimeCostMotorizedMobility = Field(
+        ...,
+        title="Travel Cost",
+        description="The travel cost of the isochrone.",
+    )
+    time_window: PTTimeWindow = Field(
+        ...,
+        title="Time Window",
+        description="The time window of the isochrone.",
+    )
+    isochrone_type: IsochroneTypePT = Field(
+        ...,
+        title="Return Type",
+        description="The return type of the isochrone.",
+    )
+
+    decay_function: IsochroneDecayFunctionPT = Field(
+        IsochroneDecayFunctionPT(),
+        title="Decay Function",
+        description="The decay function of the isochrone.",
+    )
+
+    # Defaults - not currently user configurable
+    walk_speed: float = 1.39
+    max_walk_time: int = 20
+    bike_speed: float = 4.166666666666667
+    max_bike_time: int = 20
+    bike_traffic_stress: int = 4
+    max_rides: int = 4
+    zoom: int = 9
+    percentiles: List[int] = [5]
+    monte_carlo_draws: int = 200
+
+    @property
+    def tool_type(self):
+        return ToolType.isochrone_pt
+
+    @property
+    def geofence_table(self):
+        mode = ToolType.isochrone_pt.value.replace("isochrone_", "")
+        return f"basic.geofence_{mode}"
+
+    @property
+    def input_layer_types(self):
+        return {"layer_project_id": input_layer_type_point}
+
+    @property
+    def properties_base(self):
+        return {
+            "color_range_type": ColorRangeType.sequential,
+            "color_field": {"name": "travel_cost", "type": "number"},
+            "color_scale": "quantile",
+            "breaks": self.travel_cost.steps,
+        }
+
+
+class IIsochroneCar(BaseModel):
+    """Model for the car isochrone"""
+
+    starting_points: IsochroneStartingPointsMotorizedMobility = Field(
+        ...,
+        title="Starting Points",
+        description="The starting points of the isochrone.",
+    )
+    routing_type: IsochroneRoutingTypeCar = Field(
+        ...,
+        title="Routing Type",
+        description="The routing type of the isochrone.",
+    )
+    travel_cost: IsochroneTravelTimeCostMotorizedMobility = Field(
+        ...,
+        title="Travel Cost",
+        description="The travel cost of the isochrone.",
+    )
+
+    @property
+    def tool_type(self):
+        return ToolType.isochrone_car
+
+    @property
+    def geofence_table(self):
+        mode = ToolType.isochrone_car.value.value.replace("isochrone_", "")
+        return f"basic.geofence_{mode}"
+
+    @property
+    def input_layer_types(self):
+        return {"layer_project_id": input_layer_type_point}
+
+    @property
+    def properties_base(self):
+        return {
+            "color_range_type": ColorRangeType.sequential,
+            "color_field": {"name": "travel_cost", "type": "number"},
+            "color_scale": "quantile",
+            "breaks": self.travel_cost.steps,
+        }
+
+
 class IsochroneNearbyStationAccess(BaseModel):
     """Model for the nearby stations (active mobility) isochrone"""
 
@@ -157,12 +389,12 @@ class IsochroneNearbyStationAccess(BaseModel):
         title="Starting Points",
         description="The starting points of the isochrone.",
     )
-    routing_type: RoutingActiveMobilityType = Field(
+    routing_type: IsochroneRoutingModeActiveMobility = Field(
         ...,
         title="Routing Type",
         description="The routing type of the isochrone.",
     )
-    travel_cost: TravelTimeCostActiveMobility | TravelDistanceCostActiveMobility = (
+    travel_cost: IsochroneTravelTimeCostActiveMobility | IsochroneTravelDistanceCostActiveMobility = (
         Field(
             ...,
             title="Travel Cost",
@@ -174,7 +406,7 @@ class IsochroneNearbyStationAccess(BaseModel):
         title="Scenario ID",
         description="The ID of the scenario that is used for the routing.",
     )
-    isochrone_type: IsochroneType = Field(
+    isochrone_type: IsochroneTypeActiveMobility = Field(
         ...,
         title="Return Type",
         description="The return type of the isochrone.",
@@ -208,8 +440,7 @@ class IsochroneNearbyStationAccess(BaseModel):
         }
 
 
-
-request_examples = {
+request_examples_isochrone_active_mobility = {
     "isochrone_active_mobility": {
         "single_point_walking": {
             "summary": "Single point isochrone walking",
@@ -343,4 +574,72 @@ request_examples = {
             },
         },
     }
+}
+
+
+request_examples_isochrone_pt = {
+    # 1. Isochrone for public transport with all modes
+    "all_modes_pt": {
+        "summary": "Isochrone using all PT modes",
+        "value": {
+            "starting_points": {"latitude": [52.5200], "longitude": [13.4050]},
+            "routing_type": {
+                "mode": [
+                    "bus",
+                    "tram",
+                    "rail",
+                    "subway",
+                ],
+                "egress_mode": "walk",
+                "access_mode": "walk",
+            },
+            "travel_cost": {"max_traveltime": 40, "steps": 10},
+            "time_window": {"weekday": "weekday", "from_time": 25200, "to_time": 32400},
+            "isochrone_type": "polygon",
+        },
+    },
+    # 2. Isochrone for public transport excluding bus mode
+    "exclude_bus_mode_pt": {
+        "summary": "Isochrone excluding bus mode",
+        "value": {
+            "starting_points": {"latitude": [52.5200], "longitude": [13.4050]},
+            "routing_type": {
+                "mode": [
+                    "tram",
+                    "rail",
+                    "subway",
+                ],
+                "egress_mode": "walk",
+                "access_mode": "walk",
+            },
+            "travel_cost": {"max_traveltime": 35, "steps": 5},
+            "time_window": {"weekday": "weekday", "from_time": 25200, "to_time": 32400},
+            "isochrone_type": "polygon",
+        },
+    },
+}
+
+
+request_examples_isochrone_car = {
+    # 1. Isochrone for car
+    "single_point_car": {
+        "summary": "Isochrone for a single starting point using car",
+        "value": {
+            "starting_points": {"latitude": [52.5200], "longitude": [13.4050]},
+            "routing_type": "car_peak",
+            "travel_cost": {"max_traveltime": 30, "steps": 10},
+        },
+    },
+    # 2. Multiisochrone for car
+    "multi_point_car": {
+        "summary": "Isochrone for multiple starting points using car",
+        "value": {
+            "starting_points": {
+                "latitude": [52.5200, 52.5250, 52.5300],
+                "longitude": [13.4050, 13.4150, 13.4250],
+            },
+            "routing_type": "car_peak",
+            "travel_cost": {"max_traveltime": 30, "steps": 10},
+        },
+    },
 }
